updated postprocessor 51 to work without specialzed function from histgoram.

+++ cass/postprocessing/operations.h
@@ -625,7 +625,7 @@ protected:
   shared_pointer _pHist;
 
   /** range we want to have the integral over in histogram bins */
-  std::pair<float,float> _area;
+  std::pair<size_t,size_t> _area;
 };
